+++ crates/hcl-primitives/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.3](https://github.com/martinohmann/hcl-rs/compare/hcl-primitives-v0.0.2...hcl-primitives-v0.0.3) (2023-04-21)
+
+
+### Features
+
+* implement `From` for `Cow&lt;str&gt;` <-> `InternalString` ([e352a5a](https://github.com/martinohmann/hcl-rs/commit/e352a5ac0f0eb915b0d29cc44ec2c36f5d2d9c59))
+
 ## [0.0.2](https://github.com/martinohmann/hcl-rs/compare/hcl-primitives-v0.0.1...hcl-primitives-v0.0.2) (2023-03-16)